# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shelter Center API
  version: 1.0.0
  description: "유기동물 보호소 운영 API (Animals/Adoptions/Donations/Medical/Volunteers/Stats)"
servers:
  - url: http://localhost:8080
    description: local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [timestamp, path, status, code, message]
      properties:
        timestamp: { type: string, format: date-time }
        path: { type: string }
        status: { type: integer }
        code: { type: string }
        message: { type: string }
        details: { type: object, nullable: true }
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, sort]
      properties:
        content:
          type: array
          items: {}
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        sort: { type: string }
...
    Volunteer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        note: { type: string, nullable: true }
        status:
          type: string
          enum: [PENDING, APPROVED, SUSPENDED]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
paths:
...
  /api/volunteers/volunteers:
    post:
      tags: [Volunteers]
      summary: Create volunteer application (PENDING)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 50 }
                phone: { type: string, nullable: true, maxLength: 30 }
                email: { type: string, nullable: true, maxLength: 100 }
                note: { type: string, nullable: true }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volunteer" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    get:
      tags: [Volunteers]
      summary: List volunteers (filter/search/sort/page)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, APPROVED, SUSPENDED]
        - in: query
          name: keyword
          schema: { type: string, description: "name/phone/email contains" }
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - in: query
          name: sort
          schema: { type: string, default: "createdAt,DESC" }
      responses:
        "200":
          description: paged volunteers
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/Volunteer" }
        "400":
          description: invalid query
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/volunteers/volunteers/{id}:
    get:
      tags: [Volunteers]
      summary: Get volunteer by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: volunteer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volunteer" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    patch:
      tags: [Volunteers]
      summary: Update volunteer (basic fields)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string, nullable: true }
                email: { type: string, nullable: true }
                note: { type: string, nullable: true }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volunteer" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    delete:
      tags: [Volunteers]
      summary: Delete volunteer
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: deleted
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/volunteers/volunteers/pending:
    get:
      tags: [Volunteers]
      summary: List PENDING volunteers (ADMIN planned)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: paged pending volunteers
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/Volunteer" }
  /api/volunteers/volunteers/{id}/approve:
    patch:
      tags: [Volunteers]
      summary: Approve volunteer (ADMIN planned)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volunteer" }
        "409":
          description: state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/volunteers/volunteers/{id}/suspend:
    patch:
      tags: [Volunteers]
      summary: Suspend volunteer (ADMIN planned)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: suspended
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volunteer" }
        "409":
          description: state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/volunteers/volunteers/{id}/reinstate:
    patch:
      tags: [Volunteers]
      summary: Reinstate volunteer (ADMIN planned)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: reinstated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Volunteer" }
        "409":
          description: state conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...